+++ parser_library/src/config/assembler_options.cpp
@@ -24,31 +24,33 @@ namespace hlasm_plugin::parser_library::config {
 namespace {
 using instr_set_equivalent_pair = std::pair<std::string_view, instruction_set_version>;
 
-constexpr std::array<instr_set_equivalent_pair, 23> instr_set_optable_equivalents = {
-    { { std::string_view("370"), instruction_set_version::_370 },
-        { std::string_view("DOS"), instruction_set_version::DOS },
-        { std::string_view("ESA"), instruction_set_version::ESA },
-        { std::string_view("UNI"), instruction_set_version::UNI },
-        { std::string_view("XA"), instruction_set_version::XA },
-        { std::string_view("YOP"), instruction_set_version::YOP },
-        { std::string_view("Z10"), instruction_set_version::Z10 },
-        { std::string_view("Z11"), instruction_set_version::Z11 },
-        { std::string_view("Z12"), instruction_set_version::Z12 },
-        { std::string_view("Z13"), instruction_set_version::Z13 },
-        { std::string_view("Z14"), instruction_set_version::Z14 },
-        { std::string_view("Z15"), instruction_set_version::Z15 },
-        { std::string_view("Z9"), instruction_set_version::Z9 },
-        { std::string_view("ZOP"), instruction_set_version::ZOP },
-        { std::string_view("ZS1"), instruction_set_version::ZOP },
-        { std::string_view("ZS2"), instruction_set_version::YOP },
-        { std::string_view("ZS3"), instruction_set_version::Z9 },
-        { std::string_view("ZS4"), instruction_set_version::Z10 },
-        { std::string_view("ZS5"), instruction_set_version::Z11 },
-        { std::string_view("ZS6"), instruction_set_version::Z12 },
-        { std::string_view("ZS7"), instruction_set_version::Z13 },
-        { std::string_view("ZS8"), instruction_set_version::Z14 },
-        { std::string_view("ZS9"), instruction_set_version::Z15 } }
-};
+constexpr std::array<instr_set_equivalent_pair, 25> instr_set_optable_equivalents = { {
+    { std::string_view("370"), instruction_set_version::_370 },
+    { std::string_view("DOS"), instruction_set_version::DOS },
+    { std::string_view("ESA"), instruction_set_version::ESA },
+    { std::string_view("UNI"), instruction_set_version::UNI },
+    { std::string_view("XA"), instruction_set_version::XA },
+    { std::string_view("YOP"), instruction_set_version::YOP },
+    { std::string_view("Z10"), instruction_set_version::Z10 },
+    { std::string_view("Z11"), instruction_set_version::Z11 },
+    { std::string_view("Z12"), instruction_set_version::Z12 },
+    { std::string_view("Z13"), instruction_set_version::Z13 },
+    { std::string_view("Z14"), instruction_set_version::Z14 },
+    { std::string_view("Z15"), instruction_set_version::Z15 },
+    { std::string_view("Z16"), instruction_set_version::Z16 },
+    { std::string_view("Z9"), instruction_set_version::Z9 },
+    { std::string_view("ZOP"), instruction_set_version::ZOP },
+    { std::string_view("ZS1"), instruction_set_version::ZOP },
+    { std::string_view("ZS2"), instruction_set_version::YOP },
+    { std::string_view("ZS3"), instruction_set_version::Z9 },
+    { std::string_view("ZS4"), instruction_set_version::Z10 },
+    { std::string_view("ZS5"), instruction_set_version::Z11 },
+    { std::string_view("ZS6"), instruction_set_version::Z12 },
+    { std::string_view("ZS7"), instruction_set_version::Z13 },
+    { std::string_view("ZS8"), instruction_set_version::Z14 },
+    { std::string_view("ZS9"), instruction_set_version::Z15 },
+    { std::string_view("ZSA"), instruction_set_version::Z16 },
+} };
 
 #if __cpp_lib_ranges
 static_assert(std::ranges::is_sorted(instr_set_optable_equivalents, {}, &instr_set_equivalent_pair::first));
@@ -58,13 +60,14 @@ static_assert(std::is_sorted(std::begin(instr_set_optable_equivalents),
     [](const auto& l, const auto& r) { return l.first < r.first; }));
 #endif
 
-constexpr std::array<instr_set_equivalent_pair, 54> instr_set_machine_equivalents = { {
+constexpr std::array<instr_set_equivalent_pair, 58> instr_set_machine_equivalents = { {
     { std::string_view("ARCH-0"), instruction_set_version::XA },
     { std::string_view("ARCH-1"), instruction_set_version::ESA },
     { std::string_view("ARCH-10"), instruction_set_version::Z12 },
     { std::string_view("ARCH-11"), instruction_set_version::Z13 },
     { std::string_view("ARCH-12"), instruction_set_version::Z14 },
     { std::string_view("ARCH-13"), instruction_set_version::Z15 },
+    { std::string_view("ARCH-14"), instruction_set_version::Z16 },
     { std::string_view("ARCH-2"), instruction_set_version::ESA },
     { std::string_view("ARCH-3"), instruction_set_version::ESA },
     { std::string_view("ARCH-4"), instruction_set_version::ESA },
@@ -85,6 +88,7 @@ constexpr std::array<instr_set_equivalent_pair, 54> instr_set_machine_equivalent
     { std::string_view("Z13"), instruction_set_version::Z13 },
     { std::string_view("Z14"), instruction_set_version::Z14 },
     { std::string_view("Z15"), instruction_set_version::Z15 },
+    { std::string_view("Z16"), instruction_set_version::Z16 },
     { std::string_view("Z196"), instruction_set_version::Z11 },
     { std::string_view("Z800"), instruction_set_version::ZOP },
     { std::string_view("Z890"), instruction_set_version::YOP },
@@ -95,6 +99,7 @@ constexpr std::array<instr_set_equivalent_pair, 54> instr_set_machine_equivalent
     { std::string_view("ZEC12"), instruction_set_version::Z12 },
     { std::string_view("ZS"), instruction_set_version::ZOP },
     { std::string_view("ZS-1"), instruction_set_version::ZOP },
+    { std::string_view("ZS-10"), instruction_set_version::Z16 },
     { std::string_view("ZS-2"), instruction_set_version::YOP },
     { std::string_view("ZS-3"), instruction_set_version::Z9 },
     { std::string_view("ZS-4"), instruction_set_version::Z10 },
@@ -105,6 +110,7 @@ constexpr std::array<instr_set_equivalent_pair, 54> instr_set_machine_equivalent
     { std::string_view("ZS-9"), instruction_set_version::Z15 },
     { std::string_view("ZSERIES"), instruction_set_version::ZOP },
     { std::string_view("ZSERIES-1"), instruction_set_version::ZOP },
+    { std::string_view("ZSERIES-10"), instruction_set_version::Z16 },
     { std::string_view("ZSERIES-2"), instruction_set_version::YOP },
     { std::string_view("ZSERIES-3"), instruction_set_version::Z9 },
     { std::string_view("ZSERIES-4"), instruction_set_version::Z10 },