+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## ****Unreleased****
 
+#### Added
+- Add support for z16 instructions
+- Enhance documentation for instruction mnemonics
+
+### Fixed
+
 ## [1.4.0](https://github.com/eclipse/che-che4z-lsp-for-hlasm/compare/1.3.0...1.4.0) (2022-08-30)
 
 #### Added